+++ pkg/config/validation.go
@@ -241,7 +241,7 @@ func ContractOptionsFromEnv(filePath string) (ContractsOptions, error) {
 			ParameterRegistryAddress:    config.SettlementChainParameterRegistry,
 			PayerRegistryAddress:        config.PayerRegistry,
 			PayerReportManagerAddress:   config.PayerReportManager,
-			ChainID:                     config.SettlementChainID,
+			ChainID:                     int64(config.SettlementChainID),
 			DeploymentBlock:             uint64(config.SettlementChainDeploymentBlock),
 			UnderlyingFeeToken:          config.UnderlyingFeeToken,
 			FeeToken:                    config.FeeToken,
@@ -253,7 +253,7 @@ func ContractOptionsFromEnv(filePath string) (ContractsOptions, error) {
 		AppChain: AppChainOptions{
 			GroupMessageBroadcasterAddress:   config.GroupMessageBroadcaster,
 			IdentityUpdateBroadcasterAddress: config.IdentityUpdateBroadcaster,
-			ChainID:                          config.AppChainID,
+			ChainID:                          int64(config.AppChainID),
 			MaxChainDisconnectTime:           300 * time.Second,
 			BackfillBlockPageSize:            500,
 		},
@@ -313,7 +313,7 @@ func fillConfigFromJSON(options *ContractsOptions, config *ChainConfig) {
 		options.AppChain.IdentityUpdateBroadcasterAddress = config.IdentityUpdateBroadcaster
 	}
 	if options.AppChain.ChainID == 0 || options.AppChain.ChainID == 31337 {
-		options.AppChain.ChainID = config.AppChainID
+		options.AppChain.ChainID = int64(config.AppChainID)
 	}
 	if options.AppChain.GatewayAddress == "" {
 		options.AppChain.GatewayAddress = config.AppChainGateway
@@ -339,7 +339,7 @@ func fillConfigFromJSON(options *ContractsOptions, config *ChainConfig) {
 		options.SettlementChain.PayerReportManagerAddress = config.PayerReportManager
 	}
 	if options.SettlementChain.ChainID == 0 || options.SettlementChain.ChainID == 31337 {
-		options.SettlementChain.ChainID = config.SettlementChainID
+		options.SettlementChain.ChainID = int64(config.SettlementChainID)
 	}
 	if options.SettlementChain.DeploymentBlock == 0 {
 		options.SettlementChain.DeploymentBlock = uint64(config.SettlementChainDeploymentBlock)
@@ -497,7 +497,7 @@ func validateHexAddress(address string, fieldName string, set map[string]struct{
 	}
 }
 
-func validateRPCURL(rpcURL string, chainID int, fieldName string, set map[string]struct{}) {
+func validateRPCURL(rpcURL string, chainID int64, fieldName string, set map[string]struct{}) {
 	u, err := url.Parse(rpcURL)
 	if err != nil {
 		set[fmt.Sprintf("--%s is an invalid URL, %s", fieldName, err.Error())] = struct{}{}
@@ -512,7 +512,7 @@ func validateRPCURL(rpcURL string, chainID int, fieldName string, set map[string
 	validateChainID(rpcURL, chainID, fieldName, set)
 }
 
-func validateWebsocketURL(wsURL string, chainID int, fieldName string, set map[string]struct{}) {
+func validateWebsocketURL(wsURL string, chainID int64, fieldName string, set map[string]struct{}) {
 	u, err := url.Parse(wsURL)
 	if err != nil {
 		set[fmt.Sprintf("--%s is an invalid URL, %s", fieldName, err.Error())] = struct{}{}
@@ -527,7 +527,7 @@ func validateWebsocketURL(wsURL string, chainID int, fieldName string, set map[s
 	validateChainID(wsURL, chainID, fieldName, set)
 }
 
-func validateChainID(url string, expectedChainID int, fieldName string, set map[string]struct{}) {
+func validateChainID(url string, expectedChainID int64, fieldName string, set map[string]struct{}) {
 	ctx := context.Background()
 
 	client, err := ethclient.DialContext(ctx, url)
